WIP: replace electron remote with HTTP request
It is an attempt of improving the support for large data responses.
Looks like the main problem with large data responses occur in the electron remote data transfering.
Although, even changing to use HTTP resquest it still not working. Seems like chrome lose the connection with the server when the response is to large.
My next attempt is using stream to reponse the query result.